In src: 2009-10-15  Rolf Bjarne Kvinge  <RKvinge@novell.com>
	* uri.cpp:
	* mp3.cpp:
	* runtime.h:
	* clock.cpp:
	* audio.cpp:
	* fonts.cpp:
	* timeline.cpp:
	* playlist.cpp:
	* pipeline.cpp:
	* downloader.cpp:
	* audio-alsa.cpp:
	* audio-pulse.cpp:
	* mediaplayer.cpp:
	* timemanager.cpp:
	* pipeline-asf.cpp:
	* mediaelement.cpp:
	* mms-downloader.cpp: Fix 64 bit printf format specifier on x64.

	r: andreia

In src/asf:
2009-10-15  Rolf Bjarne Kvinge  <RKvinge@novell.com>

	* asf.cpp:
	* asf-structures.cpp: Fix 64 bit printf format specifier on x64.

	r: andreia

svn path=/trunk/moon/; revision=144187

+++ src/uri.cpp
@@ -508,7 +508,7 @@ Uri::Combine (const char *relative_path)
 		if (!(filename = strrchr (path, '/')))
 			new_path = g_strdup (relative_path);
 		else
-			new_path = g_strdup_printf ("%.*s/%s", filename - path, path, g_str_has_prefix (relative_path, "../") ? relative_path+3 : relative_path);
+			new_path = g_strdup_printf ("%.*s/%s", (int) (filename - path), path, g_str_has_prefix (relative_path, "../") ? relative_path+3 : relative_path);
 		g_free (path);
 		
 		path = new_path;
